refactor: change intents.py to flags.py
