+++ interactions/client.py
@@ -14,7 +14,7 @@ from .api.gateway import WebSocket
 from .api.http import HTTPClient
 from .api.models.guild import Guild
 from .api.models.gw import Presence
-from .api.models.intents import Intents
+from .api.models.flags import Intents
 from .api.models.team import Application
 from .base import CustomFormatter, Data
 from .decor import command
